# src/main/resources/api/customerHub-openapi-v1.yml
openapi: 3.0.3
info:
  version: '1.0.0'
  title: Customer Hub APIs
  description: |
    This API serves as the backend for the Customer Hub application, providing endpoints for managing customer data.

servers:
  - url: http://localhost:8080
    description: Local server

paths:

  '/customers':
    post:
      summary: Create a new customer
      operationId: createNewCustomer
      tags:
        - 'CustomerCrud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Returns Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/400'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'

    get:
      summary: Return a list of all customers
      operationId: listCustomers
      tags:
        - 'CustomerCrud'
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Returns a list of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'

  '/customers/{customerId}':
    get:
      operationId: getCustomer
      tags:
        - 'CustomerCrud'
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        '200':
          description: Returns Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'

    put:
      summary: Update customer details
      operationId: updateCustomer
      tags:
        - 'CustomerCrud'
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Returns updated Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'

    delete:
      summary: Delete a customer
      operationId: deleteCustomer
      tags:
        - 'CustomerCrud'
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        '204':
          $ref: '#/components/responses/204'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'

  '/customers/find':
    get:
      summary: Find customers by first name and/or last name
      operationId: findCustomer
      tags:
        - 'CustomerCrud'
      parameters:
        - $ref: "#/components/parameters/FirstName"
        - $ref: "#/components/parameters/LastName"
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Returns a list of matching customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'


components:

  parameters:
    CustomerId:
      in: path
      name: customerId
      schema:
        type: string
      required: true

    FirstName:
      in: query
      name: firstName
      schema:
        type: string

    LastName:
      in: query
      name: lastName
      schema:
        type: string

    Page:
      name: page
      in: query
      description: Zero-based page index, must not be negative.
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
        maximum: 50

    Size:
      in: query
      name: size
      description: Number of items per page, must be greater than 0.
      schema:
        type: integer
        format: int32
        default: 10
        minimum: 1

  schemas:

    NotFoundError:
      type: object
      title: Not Found Error
      description: The requested customer is not found
      properties:
        message:
          type: string
          description: Error message
      required:
        - code
        - message

    ErrorBodyIsMissed:
      type: object
      title: Error Body Is Missing
      properties:
        message:
          type: string
          description: Error message

    ErrorFieldIsRequired:
      type: object
      title: Error Field Is Required
      properties:
        message:
          type: string
          description: Error message

    ErrorFieldIsInvalid:
      type: object
      title: Error Field Is Invalid
      properties:
        message:
          type: string
          description: Error message

    DuplicateError:
      type: object
      title: Duplicate record is inserted
      properties:
        message:
          type: string
          description: Error message

    CustomerRequest:
      type: object
      description: Data for creating or updating a customer. FirstName and LastName should be unique.
      required:
        - firstName
        - lastName
      properties:
        firstName:
          type: string
        lastName:
          type: string
        age:
          type: integer
        address:
          type: string
        email:
          type: string
          format: email
          example: user@example.com

    CustomerResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
        - created
      allOf:
        - $ref: '#/components/schemas/CustomerRequest'
      properties:
        id:
          type: string
        created:
          type: string
          format: date-time
          description: The creation time of the customer in UTC
        updated:
          type: string
          format: date-time

    CustomerListResponse:
      type: object
      required:
        - page
        - size
        - content
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        content:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponse'

  responses:
    204:
      description: |-
        Operation completed successfully. No content to respond.
    400:
      description: |-
        Bad request in cases of:
        * `BODY_MISSING` - Json body needs to be provided
        * `FIELD_IS_REQUIRED` - Field X is mandatory
        * `FIELD_IS_INVALID` - Field X is invalid
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorBodyIsMissed'
              - $ref: '#/components/schemas/ErrorFieldIsRequired'
              - $ref: '#/components/schemas/ErrorFieldIsInvalid'

    404:
      description: |-
        Item not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFoundError"
    422:
      description: |-
        duplicate insertion of firstName and lastName
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DuplicateError"
    500:
      description: |-
        * `TECHNICAL_ERROR` - technical error

    503:
      description: |-
        Service is unavailable